Added perspective
Not sure it is 100% correct, but at least we have perspective now.
We still have to do somthing very strange to the rotation quaternion
to make it work - which seems to be a mistake.

Also, we see the boxes inside-out, it may be related to the whole thing
being inverted because Google VR is model to cameray space, where
the jMonkeyEngine seems to be camera in world space. This needs some
investigation, or good advice from some core jME devs.